+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Drag & drop file upload #653
 - Disallow use of "any" type #624
 - Date picker #559
 - New login page #573
@@ -44,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Updated Node.js version in GitHub workflows and Dockerfile #655
 - Disallow use of any-type #624
 
   #### Added
@@ -67,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Cast integer fields as numbers #661
+- Disallow object update with invalid form #656
 - Handle required form array fields #484
 - Prevent user deleting object in published folder #486
 - Display error when replacing XML file with file with same name #483